Develop

+++ client/src/store/modules/workspaces/actions.ts
@@ -1,91 +1,62 @@
 import { WorkspaceApi } from "@/helpers/api/workspace";
-import { State, WorkspacesState, WorkspacesSortType } from "@/interfaces/store";
-import { InitializationStatus } from "@/interfaces";
+import { State, WorkspacesState, WorkspacesSortType } from "@/models/store";
+import { InitializationStatus } from "@/models";
 import { ActionTree } from "vuex";
 import axios from "axios";
 
 export const actions: ActionTree<WorkspacesState, State> = {
-  async init({ commit, rootGetters }): Promise<void> {
+  async init({ commit, rootGetters, dispatch }): Promise<void> {
+    await dispatch("user/refreshExpired", null, { root: true });
     commit("setInitStatus", { initStatus: InitializationStatus.Pending });
     try {
-      const { data } = await WorkspaceApi.getByUser(
-        rootGetters["user/accessToken"]
-      );
+      const { data } = await WorkspaceApi.getByUser(rootGetters["user/accessToken"]);
       commit("setWorkspaces", { workspaces: data.workspaces });
       commit("setInitStatus", { initStatus: InitializationStatus.Success });
     } catch (error) {
       commit("setInitStatus", { initStatus: InitializationStatus.Error });
       if (axios.isAxiosError(error)) {
-        if (error.response) {
-          if (error.response?.status === 401) {
-            // TODO: 401 handler
-            throw new Error("Access token is incorrect");
-          } else {
-            throw new Error(error.response.statusText);
-          }
-        } else {
-          throw new Error(error.message)
-        }
+        throw new Error(error.message);
       }
     }
   },
-  async create(
-    { commit, dispatch, state, rootGetters },
-    payload: { name: string }
-  ): Promise<void> {
+  async create({ commit, dispatch, state, rootGetters }, payload: { name: string }): Promise<void> {
+    await dispatch("user/refreshExpired", null, { root: true });
     try {
-      const { data } = await WorkspaceApi.create(
-        { name: payload.name },
-        rootGetters["user/accessToken"]
-      );
+      const { data } = await WorkspaceApi.create({ name: payload.name }, rootGetters["user/accessToken"]);
       commit("add", { workspace: data.workspace });
       await dispatch("sort", { sortType: state.sortType });
     } catch (error) {
       if (axios.isAxiosError(error)) {
-        if (error.response?.status === 401) {
-          // TODO: 401 handler
-          throw new Error("Access token is incorrect");
-        } else {
-          throw new Error(error.response?.statusText);
-        }
+        throw new Error(error.response?.statusText);
       }
     }
   },
-  async delete(
-    { commit, state, rootGetters },
-    payload: { id: number }
-  ): Promise<void> {
-    const workspaceIndex = state.workspaces.findIndex(
-      (item) => item.id === payload.id
-    );
+  async delete({ commit, state, dispatch, rootGetters }, payload: { id: number }): Promise<void> {
+    await dispatch("user/refreshExpired", null, { root: true });
+    const workspaceIndex = state.workspaces.findIndex((item) => item.id === payload.id);
     if (~workspaceIndex) {
       try {
-        await WorkspaceApi.delete(
-          { id: payload.id },
-          rootGetters["user/accessToken"]
-        );
+        await WorkspaceApi.delete({ id: payload.id }, rootGetters["user/accessToken"]);
         commit("delete", { workspaceIndex: workspaceIndex });
+        if (payload.id === state.currentWorkspaceId) {
+          if (state.workspaces.length > 0) {
+            commit("setCurrentId", { workspaceId: state.workspaces[0].id });
+          } else {
+            commit("setCurrentId", { workspaceId: null });
+          }
+        }
       } catch (error) {
         if (axios.isAxiosError(error)) {
-          if (error.response?.status === 401) {
-            // TODO: 401 handler
-            throw new Error("Access token is incorrect");
-          } else {
-            throw new Error(error.response?.statusText);
-          }
+          throw new Error(error.response?.statusText);
         }
       }
     } else {
       throw new Error("Workspace id is incorrect");
     }
   },
-  async update(
-    { commit, dispatch, state, rootGetters },
-    payload: { id: number; name: string }
-  ): Promise<void> {
-    const workspaceIndex = state.workspaces.findIndex(
-      (item) => item.id === payload.id
-    );
+  async update({ commit, dispatch, state, rootGetters }, payload: { id: number; name: string }): Promise<void> {
+    await dispatch("user/refreshExpired", null, { root: true });
+    const workspaceIndex = state.workspaces.findIndex((item) => item.id === payload.id);
     if (~workspaceIndex) {
       try {
         const { data } = await WorkspaceApi.update(
@@ -96,28 +67,20 @@ export const actions: ActionTree<WorkspacesState, State> = {
           oldWorkspaceIndex: workspaceIndex,
           newWorkspace: data.workspace
         });
-        if (payload.id === state.currentWorkspace?.id) {
-          commit("setCurrent", { workspace: data.workspace });
+        if (payload.id === state.currentWorkspaceId) {
+          commit("setCurrentId", { workspaceId: data.workspace.id });
         }
         await dispatch("sort", { sortType: state.sortType });
       } catch (error) {
         if (axios.isAxiosError(error)) {
-          if (error.response?.status === 401) {
-            // TODO: 401 handler
-            throw new Error("Access token is incorrect");
-          } else {
-            throw new Error(error.response?.statusText);
-          }
+          throw new Error(error.response?.statusText);
         }
       }
     } else {
       throw new Error("Workspace id is incorrect");
     }
   },
-  async sort(
-    { commit },
-    payload: { sortType: WorkspacesSortType }
-  ): Promise<void> {
+  async sort({ commit }, payload: { sortType: WorkspacesSortType }): Promise<void> {
     commit("setSortType", { sortType: payload.sortType });
     switch (payload.sortType) {
       case WorkspacesSortType.NameAscending: {
